Don't write empty non-stdout lines to console
The behavior of the test runner is that for messages that don't go through the "stdout" wrapper in mocha is that it adds a newline.  In our test setup we periodically get random empty console events which causes the test output to be staggered.  I've updated the handleConsole event handler to not add a newline.